+++ c_src/nif/impl/ed448.c.h
@@ -7,7 +7,7 @@
  * Erlang NIF functions
  */
 
-/* decaf/ed448.h */
+/* libdecaf_nif:ed448_derive_public_key/1 */
 
 static ERL_NIF_TERM
 libdecaf_nif_ed448_derive_public_key_1(ErlNifEnv *env, int argc, const ERL_NIF_TERM argv[])
@@ -25,6 +25,8 @@ libdecaf_nif_ed448_derive_public_key_1(ErlNifEnv *env, int argc, const ERL_NIF_T
 
     return out;
 }
+
+/* libdecaf_nif:ed448_sign/5 */
 
 static ERL_NIF_TERM
 libdecaf_nif_ed448_sign_5(ErlNifEnv *env, int argc, const ERL_NIF_TERM argv[])
@@ -49,6 +51,8 @@ libdecaf_nif_ed448_sign_5(ErlNifEnv *env, int argc, const ERL_NIF_TERM argv[])
 
     return out;
 }
+
+/* libdecaf_nif:ed448_sign_prehash/4 */
 
 static ERL_NIF_TERM
 libdecaf_nif_ed448_sign_prehash_4(ErlNifEnv *env, int argc, const ERL_NIF_TERM argv[])
@@ -78,6 +82,8 @@ libdecaf_nif_ed448_sign_prehash_4(ErlNifEnv *env, int argc, const ERL_NIF_TERM a
     return out;
 }
 
+/* libdecaf_nif:ed448_verify/5 */
+
 static ERL_NIF_TERM
 libdecaf_nif_ed448_verify_5(ErlNifEnv *env, int argc, const ERL_NIF_TERM argv[])
 {
@@ -101,6 +107,8 @@ libdecaf_nif_ed448_verify_5(ErlNifEnv *env, int argc, const ERL_NIF_TERM argv[])
         return ATOM_false;
     }
 }
+
+/* libdecaf_nif:ed448_verify_prehash/4 */
 
 static ERL_NIF_TERM
 libdecaf_nif_ed448_verify_prehash_4(ErlNifEnv *env, int argc, const ERL_NIF_TERM argv[])
@@ -127,4 +135,44 @@ libdecaf_nif_ed448_verify_prehash_4(ErlNifEnv *env, int argc, const ERL_NIF_TERM
         (void)decaf_ed448_prehash_destroy(hash);
         return ATOM_false;
     }
+}
+
+/* libdecaf_nif:ed448_convert_public_key_to_x448/1 */
+
+static ERL_NIF_TERM
+libdecaf_nif_ed448_convert_public_key_to_x448_1(ErlNifEnv *env, int argc, const ERL_NIF_TERM argv[])
+{
+    ErlNifBinary ed448_pubkey;
+    uint8_t *x448_pubkey = NULL;
+    ERL_NIF_TERM out_term;
+
+    if (argc != 1 || !enif_inspect_binary(env, argv[0], &ed448_pubkey) || ed448_pubkey.size != DECAF_EDDSA_448_PUBLIC_BYTES) {
+        return enif_make_badarg(env);
+    }
+
+    x448_pubkey = (uint8_t *)(enif_make_new_binary(env, DECAF_X448_PUBLIC_BYTES, &out_term));
+
+    (void)decaf_ed448_convert_public_key_to_x448(x448_pubkey, ed448_pubkey.data);
+
+    return out_term;
+}
+
+/* libdecaf_nif:ed448_convert_private_key_to_x448/1 */
+
+static ERL_NIF_TERM
+libdecaf_nif_ed448_convert_private_key_to_x448_1(ErlNifEnv *env, int argc, const ERL_NIF_TERM argv[])
+{
+    ErlNifBinary ed448_privkey;
+    uint8_t *x448_privkey = NULL;
+    ERL_NIF_TERM out_term;
+
+    if (argc != 1 || !enif_inspect_binary(env, argv[0], &ed448_privkey) || ed448_privkey.size != DECAF_EDDSA_448_PRIVATE_BYTES) {
+        return enif_make_badarg(env);
+    }
+
+    x448_privkey = (uint8_t *)(enif_make_new_binary(env, DECAF_X448_PRIVATE_BYTES, &out_term));
+
+    (void)decaf_ed448_convert_private_key_to_x448(x448_privkey, ed448_privkey.data);
+
+    return out_term;
 }