+++ c_src/nif/impl/point_255.c.h
@@ -7,7 +7,7 @@
  * Erlang NIF functions
  */
 
-/* decaf/point_255.h */
+/* libdecaf_nif:x25519_derive_public_key/1 */
 
 static ERL_NIF_TERM
 libdecaf_nif_x25519_derive_public_key_1(ErlNifEnv *env, int argc, const ERL_NIF_TERM argv[])
@@ -25,6 +25,8 @@ libdecaf_nif_x25519_derive_public_key_1(ErlNifEnv *env, int argc, const ERL_NIF_
 
     return out;
 }
+
+/* libdecaf_nif:x25519/2 */
 
 static ERL_NIF_TERM
 libdecaf_nif_x25519_2(ErlNifEnv *env, int argc, const ERL_NIF_TERM argv[])