+++ src/libdecaf_curve448.erl
@@ -17,6 +17,11 @@
 -export([eddsa_secret_to_pk/1]).
 -export([eddsa_sk_to_pk/1]).
 -export([eddsa_sk_to_secret/1]).
+-export([eddsa_keypair_to_x448_keypair/1]).
+-export([eddsa_pk_to_x448_pk/1]).
+-export([eddsa_secret_to_x448_keypair/1]).
+-export([eddsa_secret_to_x448_secret/1]).
+-export([eddsa_sk_to_x448_keypair/1]).
 % Ed448
 -export([ed448_sign/2]).
 -export([ed448_sign/3]).
@@ -71,6 +76,28 @@ eddsa_sk_to_pk(<< _:?EdDSA_SECRET_BYTES/binary, PK:?EdDSA_PK_BYTES/binary >>) ->
 eddsa_sk_to_secret(<< Secret:?EdDSA_SECRET_BYTES/binary, _:?EdDSA_PK_BYTES/binary >>) ->
 	Secret.
 
+eddsa_keypair_to_x448_keypair({<< PK:?EdDSA_PK_BYTES/binary >>, SK = << _:?EdDSA_SECRET_BYTES/binary, PK:?EdDSA_PK_BYTES/binary >>}) ->
+	eddsa_sk_to_x448_keypair(SK).
+
+eddsa_pk_to_x448_pk(<< PK:?EdDSA_PK_BYTES/binary >>) ->
+	libdecaf:ed448_convert_public_key_to_x448(PK).
+
+eddsa_secret_to_x448_keypair(<< Secret:?EdDSA_SECRET_BYTES/binary >>) ->
+	x448_keypair(eddsa_secret_to_x448_secret(Secret)).
+
+eddsa_secret_to_x448_secret(<< Secret:?EdDSA_SECRET_BYTES/binary >>) ->
+	libdecaf:ed448_convert_private_key_to_x448(Secret).
+
+eddsa_sk_to_x448_keypair(SK = << Secret:?EdDSA_SECRET_BYTES/binary, PK:?EdDSA_PK_BYTES/binary >>) ->
+	KP0 = {eddsa_pk_to_x448_pk(PK), eddsa_secret_to_x448_secret(Secret)},
+	KP1 = eddsa_secret_to_x448_keypair(Secret),
+	case KP0 =:= KP1 of
+		true ->
+			KP1;
+		false ->
+			erlang:error({badarg, [SK]})
+	end.
+
 % Ed448
 
 ed448_sign(M, << SK:?EdDSA_SK_BYTES/binary >>) when is_binary(M) ->
@@ -118,7 +145,7 @@ curve448(Scalar) when is_integer(Scalar) ->
 curve448(Scalar)
 		when is_binary(Scalar)
 		andalso byte_size(Scalar) =:= ?X448_PRIVATE_BYTES ->
-	<< U:?X448_PUBLIC_BYTES/unsigned-little-integer-unit:8 >> = libdecaf:x448_generate_key(Scalar),
+	<< U:?X448_PUBLIC_BYTES/unsigned-little-integer-unit:8 >> = libdecaf:x448_derive_public_key(Scalar),
 	U.
 
 curve448(Scalar, Base) when is_integer(Scalar) ->
@@ -136,7 +163,7 @@ curve448(Scalar, Base)
 x448(K)
 		when is_binary(K)
 		andalso byte_size(K) =:= ?X448_PRIVATE_BYTES ->
-	libdecaf:x448_generate_key(K).
+	libdecaf:x448_derive_public_key(K).
 
 x448(K, U)
 		when is_binary(K)